userdb/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
awscli==1.18.69
beautifulsoup4==4.8.2
blinker==1.4
botocore==1.16.19
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.3
djangorestframework-jwt==1.11.0
docutils==0.16
entrypoints==0.3
filelock==3.0.12
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
img2pdf==0.3.3
install==1.3.4
jmespath==0.9.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pdfarranger==1.5.3
pexpect==4.6.0
pikepdf==1.10.3+dfsg
Pillow==7.0.0
protobuf==3.6.1
pyasn1==0.4.2
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.4.1
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.4.2
wadllib==1.3.3
webencodings==0.5.1
xkit==0.0.0
